requirements.txt
numpy
pybind11
ultralytics
python-socketio
aiohttp
influxdb-client
